Moved find and thread_find funcs in separate files
The `find`, `vect_find` and `thread_find` functions have seen their
implementations moved to `find.c` and `thread_find.c` respectively.

This leads to a cleaner `simdbmk.c` (which has actually been renamed
`main.c` in the same commit since it contains nothing but the `main()`
function now) and enables us to put a ****load of comments in `find.h`
and `thread_find.h`.
