+++ find.h
@@ -1,0 +1,53 @@
+/*
+ * ============================================================================
+ *
+ *       Filename:  find.h
+ *
+ *    Description:  The header file for the find function and it's
+ *                  SIMD-optimized counterpart: vect_find.
+ *
+ *        Version:  1.0
+ *        Created:  09/01/2016 19:54:25
+ *       Revision:  none
+ *       Compiler:  gcc
+ *
+ *        Authors:  Etienne LAFARGE (etienne.lafarge@mines-paristech.fr),
+ *                  Vincent Villet (vincent.villet@mines-paristech.fr)
+ *
+ *   Organization:  École Nationale Supérieure des Mines de Paris
+ *
+ * ============================================================================
+ */
+
+#ifndef _FIND_H_
+#define _FIND_H_
+
+/**
+ * Looks for val in U between the indexes i_start and i_end and jumping by
+ * i_step at a time. It will return the number of found occurences of val and
+ * put their positions in the pointer **ind_val.
+ *
+ * Why is **ind_val a pointer on a pointer of ints ? Well that's pretty simple:
+ * what we need to do is populate an array, which is represented by a pointer
+ * (*int_val) and a size (called c and returned by the function find). At first
+ * you'd think that just passing a pointer and make it change would be fine,
+ * but the point is on a function call, the arguments are copied in memory.
+ * Therefore, it's that copy of the pointer that will be modified by find, not
+ * the pointer itself. That's why we need to get a pointer pointing on that
+ * pointer: therefore the pointer gives us the address of the pointer pointing
+ * at the beginning of the array and we we can use this "address" to actually
+ * modify the array every time we call realloc.
+ *
+ * In short the answer is: "because arguments get copied when a function is
+ * called so we need to use pointers, and since the argument itself is a
+ * pointer, we need a pointer on a pointer".
+ */
+int find(int *U, int i_start, int i_end, int i_step, int val, int **ind_val);
+
+
+
+int vect_find(int *U, int i_start, int i_end, int i_step, int val,
+              int **ind_val);
+
+
+#endif
