+++ utilities.c
@@ -1,5 +1,5 @@
 /*
- * =====================================================================================
+ * ============================================================================
  *
  *       Filename:  utilities.c
  *
@@ -17,7 +17,7 @@
  *
  *   Organization:  École Nationale Supérieure des Mines de Paris
  *
- * =====================================================================================
+ * ============================================================================
  */
 #include "utilities.h"
